--- .project
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<projectDescription>
-	<name>ExternalConfigReload</name>
-	<comment></comment>
-	<projects>
-	</projects>
-	<buildSpec>
-		<buildCommand>
-			<name>org.eclipse.jdt.core.javabuilder</name>
-			<arguments>
-			</arguments>
-		</buildCommand>
-	</buildSpec>
-	<natures>
-		<nature>com.springsource.sts.grails.core.nature</nature>
-		<nature>org.eclipse.jdt.groovy.core.groovyNature</nature>
-		<nature>org.eclipse.jdt.core.javanature</nature>
-	</natures>
-	<linkedResources>
-		<link>
-			<name>.link_to_grails_plugins</name>
-			<type>2</type>
-			<locationURI>GRAILS_ROOT/1.3.7/projects/ExternalConfigReload/plugins</locationURI>
-		</link>
-	</linkedResources>
-</projectDescription>
