--- .settings/com.springsource.sts.grails.core.prefs
@@ -1,4 +1,0 @@
-#Thu Aug 25 11:14:15 MDT 2011
-com.springsource.sts.grails.core.com.springsource.sts.grails.core.install.name=Grails 1.3.7
-com.springsource.sts.grails.core.use.default.install=false
-eclipse.preferences.version=1
