--- .settings/org.codehaus.groovy.eclipse.preferences.prefs
@@ -1,3 +1,0 @@
-#Created by grails
-eclipse.preferences.version=1
-groovy.dont.generate.class.files=true
