--- application.properties
@@ -1,6 +1,0 @@
-#Grails Metadata file
-#Tue Oct 30 15:47:26 MDT 2012
-app.grails.version=2.2.2
-app.name=ExternalConfigReload
-plugins.hibernate=2.2.2
-plugins.tomcat=2.2.2
